Moves get_match_for_seq func to utils
